servoshell: Move panic hook into separate file
